requirements.txt
reflex==0.6.4
python-decouple
openai
